src/requirements.txt
flask==3.0.0
flask-login==0.6.3
flask-wtf==1.2.1
flask-limiter==3.5.0
flask-sqlalchemy>=3.1.1
sqlalchemy>=2.0.23
alembic>=1.13.1
pydantic>=2.10.0
email-validator>=2.0.0
numpy>=1.26.0,<2.0.0
pandas>=2.2.0
matplotlib>=3.8.0
reportlab==4.0.7
cryptography>=41.0.0
bcrypt>=4.0.0
pillow>=10.0.0
google-genai>=0.1.0
anthropic>=0.42.0
